Rename of packages with Java Name Convention and rename Subject class and Object class
