--- OBL/src/InstructionObject/InstructionValidator.java
@@ -1,14 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package InstructionObject;
-
-/**
- *
- * @author Usuario
- */
-public abstract class InstructionValidator {
-    public abstract Boolean validate(String instruction);
-}
