--- OBL/src/Labeling/SecurityLevel.java
@@ -1,11 +1,0 @@
-package Labeling;
-
-/**
- *
- * @author martellocauda
- */
-public enum SecurityLevel {
-    LOW,
-    MEDIUM,
-    HIGH
-}
