--- OBL/src/ObjectManager/Exceptions/ObjectException.java
@@ -1,16 +1,0 @@
-package ObjectManager.Exceptions;
-
-/**
- *
- * @author martellocauda
- */
-public class ObjectException extends Exception {
-
-    public ObjectException(String message) {
-        super(message);
-    }
-
-    public ObjectException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
