--- OBL/src/Persistence/Exceptions/PersistenceFileException.java
@@ -1,20 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Persistence.Exceptions;
-
-/**
- *
- * @author Usuario
- */
-public class PersistenceFileException extends Exception {
-    public PersistenceFileException(String message) {
-        super(message);
-    }
-
-    public PersistenceFileException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
