--- OBL/src/Persistence/Persistence.java
@@ -1,13 +1,0 @@
-package Persistence;
-
-import java.io.IOException;
-
-/**
- *
- * @author Usuario
- */
-public interface Persistence {
-    public Boolean openPersistence(String stringPersistence) throws IOException;
-    public String retrieveOneLine() throws IOException;
-    public Boolean closePersistence() throws IOException;
-}
