--- OBL/src/ReferenceMonitor/Exceptions/ReferenceMonitorException.java
@@ -1,16 +1,0 @@
-package ReferenceMonitor.Exceptions;
-
-/**
- *
- * @author martellocauda
- */
-public class ReferenceMonitorException extends Exception {
-
-    public ReferenceMonitorException(String message) {
-        super(message);
-    }
-
-    public ReferenceMonitorException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
