--- OBL/src/Subject/Exceptions/SubjectException.java
@@ -1,16 +1,0 @@
-package Subject.Exceptions;
-
-/**
- *
- * @author martellocauda
- */
-public class SubjectException extends Exception {
-
-    public SubjectException(String message) {
-        super(message);
-    }
-
-    public SubjectException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
