+++ OBL/src/entity/obl_subject/exception/SubjectException.java
@@ -1,0 +1,16 @@
+package entity.obl_subject.exception;
+
+/**
+ *
+ * @author martellocauda
+ */
+public class SubjectException extends Exception {
+
+    public SubjectException(String message) {
+        super(message);
+    }
+
+    public SubjectException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
