+++ OBL/src/labeling/SecurityLevel.java
@@ -1,0 +1,11 @@
+package labeling;
+
+/**
+ *
+ * @author martellocauda
+ */
+public enum SecurityLevel {
+    LOW,
+    MEDIUM,
+    HIGH
+}
