+++ OBL/src/object_manager/exception/ObjectManagerException.java
@@ -1,0 +1,16 @@
+package object_manager.exception;
+
+/**
+ *
+ * @author martellocauda
+ */
+public class ObjectManagerException extends Exception {
+
+    public ObjectManagerException(String message) {
+        super(message);
+    }
+
+    public ObjectManagerException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
