+++ OBL/src/reference_monitor/exception/ReferenceMonitorException.java
@@ -1,0 +1,16 @@
+package reference_monitor.exception;
+
+/**
+ *
+ * @author martellocauda
+ */
+public class ReferenceMonitorException extends Exception {
+
+    public ReferenceMonitorException(String message) {
+        super(message);
+    }
+
+    public ReferenceMonitorException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
